Complete Run

Run an elevation file through all functions to calculate watershed flow and fill patterns.

complete_run(file, nrow = NULL, ncol = NULL, missing_value = -9999,
  max_area = 10, max_depth = 0.5, folder_out = NULL, clean = FALSE,
  clim = NULL, rlim = NULL, continue = NULL, end = NULL, log = TRUE,
  report = TRUE, verbose = FALSE, quiet = FALSE)

Arguments

file
Character. Elevation file (see load_file for supported file types.

nrow
Numeric. Number of rows in dem file (required for dbf files with a single column, but can be automatically assessed from files with x and y coordinates.

ncol
Numeric. Number of columns in dem file (required for dbf files with a single column, but can be automatically assessed from files with x and y coordinates.

missing_value
Numeric/Character. Symbols which define missing data

max_area
Numeric. Largest area of pits to be removed during initial pit removal

max_depth
Numeric. Largest depth of pits to be removed during initial pit removal

folder_out
Charater. Folder to store output files. Defaults to location of dem file if not specified

clean
Logical. Remove all backup files and output files from previous runs in this folder?

clim
Numeric vector. Column limits if specifying a subset of the dem

rlim
Numeric vector. Row limits if specifying a subset of the dem

continue
Character. If resuming a run, where to resume (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

log
Logical. Create log file recording progress?

report
Logical. Create html report of results?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Details

For information regarding loading other file types see load_file.

For resuming or ending a run, continue or end must be one of the following:

directions (Calculating Directions)

watersheds (Calculating Watersheds)

local (Initial Pit Removal)

pond (Calculating Pond Shed Statistics - Second Pit Removal)

fill (Calculating Fill Shed Statistics - Third Pit Removal)

inverted (Calculating Directions on Inverted DEM)

iwatersheds (Calculating Inverted Watersheds)

ilocal (Initial Inverted Pit Removal)

report (Create the final report)

Examples

# NOT RUN {
# Basic Run

# (Find test files in extdata folder)

complete_run(file = "testELEV.dbf", nrow = 150, ncol = 150)

# Specify parameters for initial pit removal
complete_run(file = "testELEV.dbf", nrow = 150, ncol = 150,
             max_area = 5, max_depth = 2)
# }