Map flow through the landscape

Run an elevation file through all functions to calculate watershed flow and fill patterns. Based on FlowMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.

flow_mapper(file, nrow = NULL, ncol = NULL, missing_value = -9999,
  max_area = 10, max_depth = 0.5, out_folder = NULL,
  out_format = "rds", clean = FALSE, clim = NULL, rlim = NULL,
  resume = NULL, end = NULL, log = TRUE, report = TRUE,
  verbose = FALSE, quiet = FALSE)

Arguments

file
Character. Elevation file (see load_file for supported file types.

nrow
Numeric. Number of rows in dem file (required for dbf files with a single column, but can be automatically assessed from files with x and y coordinates.

ncol
Numeric. Number of columns in dem file (required for dbf files with a single column, but can be automatically assessed from files with x and y coordinates.

missing_value
Numeric/Character. Symbols which define missing data

max_area
Numeric. Largest area of pits to be removed during initial pit removal

max_depth
Numeric. Largest depth of pits to be removed during initial pit removal

out_folder
Character. Folder in which to store output files. Defaults to folder in the same location and with the same name as the dem file

clean
Logical. Remove all backup files and output files from previous runs in this folder?

clim
Numeric vector. Column limits if specifying a subset of the dem

rlim
Numeric vector. Row limits if specifying a subset of the dem

resume
Character. From which stage should the run be resumed? (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

log
Logical. Create log file recording progress?

report
Logical. Create html report of results?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Details

For information regarding loading other file types see load_file.

For resuming or ending a run, resume or end must be one of the following:

directions (Calculating Directions)

watersheds (Calculating Watersheds)

local (Initial Pit Removal)

pond (Calculating Pond Shed Statistics - Second Pit Removal)

fill (Calculating Fill Shed Statistics - Third Pit Removal)

inverted (Calculating Directions on Inverted DEM)

iwatersheds (Calculating Inverted Watersheds)

ilocal (Initial Inverted Pit Removal)

report (Create the final report)

Examples

# Basic Run
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 150, ncol = 150)

#> CALCULATING DIRECTIONS

#> CALCULATING WATERSHEDS

#> REMOVING INITIAL PITS

#> CALCULATING POND (GLOBAL) WATERSHEDS

#> CALCULATING FILL PATTERNS

#> INVERTING DEM

#> CALCULATING INVERTED DIRECTIONS

#> CALCULATING INVERTED WATERSHEDS

#> REMOVING INVERTED INITIAL PITS

#> SAVING OUTPUT

#> CREATING REPORT

#> Run took: 0.38 min

# Specify parameters for initial pit removal
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
            out_folder = "./testELEV/", nrow = 150, ncol = 150,
            max_area = 5, max_depth = 2)

#> CALCULATING DIRECTIONS

#> CALCULATING WATERSHEDS

#> REMOVING INITIAL PITS

#> CALCULATING POND (GLOBAL) WATERSHEDS

#> CALCULATING FILL PATTERNS

#> INVERTING DEM

#> CALCULATING INVERTED DIRECTIONS

#> CALCULATING INVERTED WATERSHEDS

#> REMOVING INVERTED INITIAL PITS

#> SAVING OUTPUT

#> CREATING REPORT

#> Run took: 0.37 min

# Clean up (remove created folder and output)
unlink("./testELEV/", recursive = TRUE)