Create plots of water flow

Plots water flow and watersheds. See the flow_plot article/vignette for examples.

flow_plot(
  db,
  type = "relief",
  dir = FALSE,
  seqno = FALSE,
  highlight = FALSE,
  shed = FALSE,
  shed_type = "shedno",
  pits = FALSE,
  upslope_threshold = NULL,
  cells = NULL,
  clim = NULL,
  rlim = NULL,
  stats = NULL,
  missing = NA
)

Arguments

db
Data frame. Cell by cell data on the elevation of the watershed. Output by LITAP's flow_mapper() function.

type
Character. Either relief or elevation. Defaults to relief.

dir
Logical. Include flow directions?

seqno
Logical. Include cell numbering?

highlight
Logical. Highlight selected cells?

shed
Logical. Show watersheds?

shed_type
Character. Which type of watershed, must be included as a column in the data frame. Can be one of 'initial', 'local', 'pond', or 'fill'.

pits
Logical. Show watershed pits (lowest point)

upslope_threshold
Numeric. If dir = TRUE, only show flow directions for cells with >= this many cells which drain to it.

cells
Vector. Which cells to show

clim
Numeric vector. Column limits in format c(0, 100)

rlim
Numeric vector. Row limits in format c(0, 100)

stats
Data frame. Data frame of watershed stats to highlight pour points.

missing
Character. What is the value of missing data? Defaults to NA