Test DEM

A small data frame for testing and running examples from helper functions. This is how DEMs look once imported by LITAP.

test_dem

Format

A data frame with 35154 rows and 6 variables:

seqno

Cell number

elev

Cell elevation

x

Longitudinal value

y

Latitudinal value

row

Cell row

col

Cell column

missing

Missing?

buffer

Buffer cell? Buffers are padding added to edges of a dem