Calculate spatial entities required for WEPP

This function takes backup data frame output from flow_mapper() and form_mapper() calculates hillslope, channel segments and impoundment spatial entities required for input in to WEPP (Water Erosion and Prediction Project) (among other metrics). Based on FacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.

wepp_mapper(
  folder,
  grid,
  chan_length = 200,
  upslope_threshold = 300,
  out_format = "rds",
  clean = FALSE,
  resume = NULL,
  end = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE
)

Arguments

folder
Character. Location of flow_mapper() output

grid
Numeric. Grid size for the original dem

chan_length
Numeric. Channel length maximum length. Used to split channels into segments

upslope_threshold
Numeric. Threshold of upslope cells to define channel cells. #' Cells with an upslope value larger than this are considered channel cells

resume
Character. From which stage should the run be resumed? (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

log
Logical. Create log file recording progress?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Examples

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 150, ncol = 150)

#> CALCULATING DIRECTIONS

#> CALCULATING WATERSHEDS

#> REMOVING INITIAL PITS

#> CALCULATING POND (GLOBAL) WATERSHEDS

#> CALCULATING FILL PATTERNS

#> INVERTING DEM

#> CALCULATING INVERTED DIRECTIONS

#> CALCULATING INVERTED WATERSHEDS

#> REMOVING INVERTED INITIAL PITS

#> SAVING OUTPUT

#> CREATING REPORT

#> Run took: 0.43 min

# Now can run wepp_mapper()
wepp_mapper(folder = "./testELEV/")

#> Error in write_log("  Input folder = ", normalizePath(folder), "\n", "  grid =  ",     grid, "\n", "  chan_length = ", chan_length, "\n", "  upslope_threshold = ",     upslope_threshold, "\n", log = log_file): argument "grid" is missing, with no default