Common arguments for flow_mapper(), form_mapper(), facet_mapper(), a nd wepp_mapper()

Common arguments for flow_mapper(), form_mapper(), facet_mapper(), a nd wepp_mapper()

Arguments

grid
Numeric. Grid size in m of the input DEM file

resume
Character. From which stage should the run be resumed? (see

clean
Logical. Remove all output files from previous runs in this folder?

report
Logical. Create html report of results?

log
Logical. Create log file recording progress?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

debug
Logical. If TRUE, output files contain intermediate columns useful for debugging (e.g., 'buffer', 'seqno_buffer', etc.) Default FALSE.