Complete Run

Run an elevation dem file through all functions to caculate watershed flow and fill patterns

complete_run(file, nrow, ncol, missing_value = -9999, max_area = 10,
  max_depth = 0.5, folder_out = NULL, clean = FALSE, xlim = NULL,
  ylim = NULL, continue = NULL, end = NULL, verbose = FALSE,
  quiet = FALSE)

Arguments

file
Character. Elev.dbf dem file

nrow
Numeric. Number of rows in dem file

ncol
Numeric. Number of columns in dem file

missing_value
Numeric/Character. Symbols which define missing data

max_area
Numeric. Largest area of pits to be removed during initial pit removal

max_depth
Numeric. Largest depth of pits to be removed during initial pit removal

folder_out
Charater. Folder to store output files. Defaults to location of dem file if not specified

clean
Logical. Remove all backup files and output files from previous runs?

xlim
Numeric vector. Column limits if specifying a subset of the dem

ylim
Numeric vector. Row limits if specifying a subset of the dem

continue
Character. If resuming a run, where to resume (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Details

For resuming or ending a run, continue or end must be one of the following:

directions (Calculating Directions)

watersheds (Calculating Watersheds)

watershed_area (Calculating Watershed Area

local (Initial Pit Removal)

pond (Calculating Pond Shed Statistics - Second Pit Removal)

fill (Calculating Fill Shed Statistics - Third Pit Removal)

inverted (Calculating Directions on Inverted DEM)

iwatersheds (Calculating Inverted Watersheds)

iwatershed_area (Calculating Watershed Area)

ilocal (Initial Inverted Pit Removal)

Examples

# Basic Run
not_run({
  complete_run(file = "testElev.dbf", nrow = 100, ncol = 100)

  # Specify parameters for initial pit removal
  complete_run(file = "testElev.dbf", nrow = 100, ncol = 100,
               max_area = 5, max_depth = 2)
})