Calculate landform facets using LSM logic

This function takes backup data frame output from flow_mapper() and form_mapper() calculates fuzzy attributes (among other metrics). Based on FacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.

facet_mapper(
  folder,
  arule = NULL,
  crule,
  n_remove = 9,
  procedure = "lsm",
  zone = NULL,
  out_format = "rds",
  clean = FALSE,
  resume = NULL,
  end = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE
)

Arguments

folder
Character. Location of flow_mapper() output

arule
Character. Location of ARULE file. If NULL, A Rules are derived from the dem file (see Details).

crule
Character. Location of CRULE file

n_remove
Numeric. Number of cells (rows/columns) to remove around the edge of the dem before deriving the A Rules.

procedure
Character. Which LSM procedure to use. One of lsm (Original LandMapR program) or bc_pem (newer BC-PEM Direct-to-Site-SEries DSS program).

zone
file. If procedure = "bc_pem", zones must either be defined for each seqno in the weti dem file, OR must be provided as an index file here. With a zone defined for each seqno. zone file can be either dem (.dem), Excel (.xlsx, .xls), or text (.txt, .csv, .dat)

out_format
Character. What format should the data be output as? "csv" for Comma-separated values, "dbf" for dbf database files, or "rds" for R data format.

clean
Logical. Remove all output files from previous runs in this folder?

resume
Character. From which stage should the run be resumed? (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

log
Logical. Create log file recording progress?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Details

Based on the technique described in Li et al. 2011, if no arule file is provided, the ARULE cutoffs are calulated from the form_mapper() dem files. These A Rules are saved as afile_derived.csv in the folder provided.

Procedure lsm refers to... Procedure bc_pem refers to...

For resuming or ending a run, resume or end must be one of the following:

attributes

classes

@references Sheng Li, David A. Lobb, Brian G. McConkey, R. A. MacMillan, Alan Moulin, and Walter R. Fraser. 2011. Extracting topographic characteristics of landforms typical of Canadian agricultural landscapes for agri-environmental modeling. I. Methodology. Canadian Journal of Soil Science 91(2), 251-266. https://doi.org/10.1139/CJSS10080

Examples

facet_mapper,Calculate landform facets using LSM logic,Calculate landform facets using LSM logic — facet_mapper,Source: R/facet_mapper.R,facet_mapper.Rd,character(0),facet_mapper,character(0),list(title = "Description", contents = "
This function takes backup data frame output from flow_mapper() and\nform_mapper() calculates fuzzy attributes (among other metrics).  Based on\nFacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.
"),list(description = "This function takes backup data frame output from flow_mapper() and\nform_mapper() calculates fuzzy attributes (among other metrics).  Based on\nFacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions."),facet_mapper(
  folder,
  arule = NULL,
  crule,
  n_remove = 9,
  procedure = "lsm",
  zone = NULL,
  out_format = "rds",
  clean = FALSE,
  resume = NULL,
  end = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE
),list(list(name = "folder", description = "
Character. Location of flow_mapper() output
"), list(name = "arule", description = "
Character. Location of ARULE file. If NULL, A Rules are derived\nfrom the dem file (see Details).
"), list(name = "crule", description = "
Character. Location of CRULE file
"), list(name = "n_remove", description = "
Numeric. Number of cells (rows/columns) to remove around the\nedge of the dem before deriving the A Rules.
"), 
    list(name = "procedure", description = "
Character. Which LSM procedure to use. One of lsm\n(Original LandMapR program) or bc_pem (newer BC-PEM Direct-to-Site-SEries\nDSS program).
"), list(name = "zone", description = "
file. If procedure = \"bc_pem\", zones must either be defined for\neach seqno in the weti dem file, OR must be provided as an index file here.\nWith a zone defined for each seqno. zone file can be either dem (.dem),\nExcel (.xlsx, .xls), or text (.txt, .csv, .dat)
"), 
    list(name = "out_format", description = "
Character. What format should the data be output as? \"csv\"\nfor Comma-separated values, \"dbf\" for dbf database files, or \"rds\" for R\ndata format.
"), list(name = "clean", description = "
Logical. Remove all output files from previous runs in this\nfolder?
"), list(name = "resume", description = "
Character. From which stage should the run be resumed? (see\nDetails below)
"), list(name = "end", description = "
Character. If ending a run after a particular step, which step\n(see Details below)
"), 
    list(name = "log", description = "
Logical. Create log file recording progress?
"), list(name = "verbose", description = "
Logical. Output extra progress messages.
"), list(name = "quiet", description = "
Logical. Suppress all messages.
")),TRUE,

# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90)

#> CALCULATING DIRECTIONS

#> CALCULATING WATERSHEDS

#> REMOVING INITIAL PITS

#> CALCULATING POND (GLOBAL) WATERSHEDS

#> CALCULATING FILL PATTERNS

#> INVERTING DEM

#> CALCULATING INVERTED DIRECTIONS

#> CALCULATING INVERTED WATERSHEDS

#> REMOVING INVERTED INITIAL PITS

#> SAVING OUTPUT

#> CREATING REPORT

#> Run took: 0.2 min

# And form_mapper()
form_mapper(folder = "./testELEV/", grid = 5)

#> SETUP

#> CALCULATING FORM

#> CALCULATING WETNESS INDICES

#> CALCULATING RELIEF DERIVITIVES

#> CALCULATING SLOPE LENGTH

#> Run took: 0.1 min

# Now can run facet_mapper() - Derive A Rules
crule <- system.file("extdata", "crule.dbf", package = "LITAP")
facet_mapper(folder = "./testELEV/", arule = NULL, crule = crule)

#> Formatting arule file

#>   - Renaming 'slope' to 'slope_pct'

#> Formatting crule file

#> 

#> CALCULATING FUZZY ATTRIBUTES

#> CALCULATING CLASSES

#> Run took: 0 min

# Now can run facet_mapper() - supply A Rules
arule <- system.file("extdata", "arule.dbf", package = "LITAP")
crule <- system.file("extdata", "crule.dbf", package = "LITAP")
facet_mapper(folder = "./testELEV/", arule = arule, crule = crule)

#> Formatting arule file

#>   - Renaming 'slope' to 'slope_pct'

#> Formatting crule file

#> 

#> CALCULATING FUZZY ATTRIBUTES

#> CALCULATING CLASSES

#> Run took: 0 min

# Clean up (remove all output)
unlink("./testELEV/", recursive = TRUE)

,list(list(title = "Details", contents = "
Based on the technique described in Li et al. 2011, if no arule file is\nprovided, the ARULE cutoffs are calulated from the form_mapper() dem\nfiles. These A Rules are saved as afile_derived.csv in the folder\nprovided.
\n
Procedure lsm refers to... Procedure bc_pem refers to...
\n
For resuming or ending a run, resume or end must be\none of the following:

\n

attributes

\n

classes

\n
\n\n
@references\nSheng Li, David A. Lobb, Brian G. McConkey, R. A. MacMillan, Alan Moulin,\nand Walter R. Fraser. 2011. Extracting topographic characteristics of landforms\ntypical of Canadian agricultural landscapes for agri-environmental modeling.\nI. Methodology. Canadian Journal of Soil Science 91(2), 251-266.\nhttps://doi.org/10.1139/CJSS10080
", 
    slug = "details")),2021,list(name = "LITAP", version = "0.4.1.9000", authors = "Steffi LaZerte, Sheng Li, Agriculture and Agri-Food Canada"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "LITAP"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Articles\n
\n
\n  News\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE