Map flow through the landscape

Run an elevation file through all functions to calculate watershed flow and fill patterns. Based on FlowMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.

flow_mapper(
  file,
  nrow = NULL,
  ncol = NULL,
  missing_value = -9999,
  max_area = 10,
  max_depth = 0.5,
  out_folder = NULL,
  out_format = "rds",
  clean = FALSE,
  clim = NULL,
  rlim = NULL,
  resume = NULL,
  end = NULL,
  log = TRUE,
  report = TRUE,
  verbose = FALSE,
  quiet = FALSE
)

Arguments

file
Character. Elevation file (see load_file) for supported file types.

nrow
Numeric. Number of rows in dem file (required for dbf files with a single column, but can be automatically assessed from files with x and y coordinates.

ncol
Numeric. Number of columns in dem file (required for dbf files with a single column, but can be automatically assessed from files with x and y coordinates.

missing_value
Numeric/Character. Symbols which define missing data

max_area
Numeric. Largest area of pits to be removed during initial pit removal

max_depth
Numeric. Largest depth of pits to be removed during initial pit removal

out_folder
Character. Folder in which to store output files. Defaults to folder in the same location and with the same name as the dem file

out_format
Character. What format should the data be output as? "csv" for Comma-separated values, "dbf" for dbf database files, or "rds" for R data format.

clean
Logical. Remove all backup files and output files from previous runs in this folder?

clim
Numeric vector. Column limits if specifying a subset of the dem

rlim
Numeric vector. Row limits if specifying a subset of the dem

resume
Character. From which stage should the run be resumed? (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

log
Logical. Create log file recording progress?

report
Logical. Create html report of results?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Details

For information regarding loading other file types see load_file.

For resuming or ending a run, resume or end must be one of the following:

directions (Calculating Directions)

watersheds (Calculating Watersheds)

local (Initial Pit Removal)

pond (Calculating Pond Shed Statistics - Second Pit Removal)

fill (Calculating Fill Shed Statistics - Third Pit Removal)

slope (Slope Gradient and Curvature values)

inverted (Calculating Directions on Inverted DEM)

iwatersheds (Calculating Inverted Watersheds)

ilocal (Initial Inverted Pit Removal)

report (Create the final report)

Examples

flow_mapper,Map flow through the landscape,Map flow through the landscape — flow_mapper,Source: R/flow_mapper.R,flow_mapper.Rd,character(0),flow_mapper,character(0),list(title = "Description", contents = "
Run an elevation file through all functions to calculate watershed flow and\nfill patterns. Based on FlowMapR by R. A. (Bob) MacMillan, LandMapper\nEnvironmental Solutions.
"),list(description = "Run an elevation file through all functions to calculate watershed flow and\nfill patterns. Based on FlowMapR by R. A. (Bob) MacMillan, LandMapper\nEnvironmental Solutions."),flow_mapper(
  file,
  nrow = NULL,
  ncol = NULL,
  missing_value = -9999,
  max_area = 10,
  max_depth = 0.5,
  out_folder = NULL,
  out_format = "rds",
  clean = FALSE,
  clim = NULL,
  rlim = NULL,
  resume = NULL,
  end = NULL,
  log = TRUE,
  report = TRUE,
  verbose = FALSE,
  quiet = FALSE
),list(list(name = "file", description = "
Character. Elevation file (see load_file) for\nsupported file types.
"), list(name = "nrow", description = "
Numeric. Number of rows in dem file (required for dbf files with\na single column, but can be automatically assessed from files with x and y\ncoordinates.
"), list(name = "ncol", description = "
Numeric. Number of columns in dem file (required for dbf files\nwith a single column, but can be automatically assessed from files with x\nand y coordinates.
"), 
    list(name = "missing_value", description = "
Numeric/Character. Symbols which define missing data
"), list(name = "max_area", description = "
Numeric. Largest area of pits to be removed during initial\npit removal
"), list(name = "max_depth", description = "
Numeric. Largest depth of pits to be removed during initial\npit removal
"), list(name = "out_folder", description = "
Character. Folder in which to store output files. Defaults\nto folder in the same location and with the same name as the dem file
"), 
    list(name = "out_format", description = "
Character. What format should the data be output as? \"csv\"\nfor Comma-separated values, \"dbf\" for dbf database files, or \"rds\" for R\ndata format.
"), list(name = "clean", description = "
Logical. Remove all backup files and output files from previous\nruns in this folder?
"), list(name = "clim", description = "
Numeric vector. Column limits if specifying a subset of the dem
"), list(name = "rlim", description = "
Numeric vector. Row limits if specifying a subset of the dem
"), 
    list(name = "resume", description = "
Character. From which stage should the run be resumed? (see\nDetails below)
"), list(name = "end", description = "
Character. If ending a run after a particular step, which step\n(see Details below)
"), list(name = "log", description = "
Logical. Create log file recording progress?
"), list(name = "report", description = "
Logical. Create html report of results?
"), list(name = "verbose", description = "
Logical. Output extra progress messages.
"), 
    list(name = "quiet", description = "
Logical. Suppress all messages.
")),TRUE,
# Basic Run
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90)

#> CALCULATING DIRECTIONS

#> CALCULATING WATERSHEDS

#> REMOVING INITIAL PITS

#> CALCULATING POND (GLOBAL) WATERSHEDS

#> CALCULATING FILL PATTERNS

#> INVERTING DEM

#> CALCULATING INVERTED DIRECTIONS

#> CALCULATING INVERTED WATERSHEDS

#> REMOVING INVERTED INITIAL PITS

#> SAVING OUTPUT

#> CREATING REPORT

#> Run took: 0.17 min

# Specify parameters for initial pit removal
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
            out_folder = "./testELEV/", nrow = 90, ncol = 90,
            max_area = 5, max_depth = 2)

#> CALCULATING DIRECTIONS

#> CALCULATING WATERSHEDS

#> REMOVING INITIAL PITS

#> CALCULATING POND (GLOBAL) WATERSHEDS

#> CALCULATING FILL PATTERNS

#> INVERTING DEM

#> CALCULATING INVERTED DIRECTIONS

#> CALCULATING INVERTED WATERSHEDS

#> REMOVING INVERTED INITIAL PITS

#> SAVING OUTPUT

#> CREATING REPORT

#> Run took: 0.17 min

# Clean up (remove created folder and output)
unlink("./testELEV/", recursive = TRUE)

,list(list(title = "Details", contents = "
For information regarding loading other file types see\nload_file.
\n
For resuming or ending a run, resume or end must be one of\nthe following:
\n
\n

directions (Calculating Directions)

\n

watersheds (Calculating Watersheds)

\n

local (Initial Pit Removal)

\n

pond (Calculating Pond Shed Statistics - Second Pit Removal)

\n

fill (Calculating Fill Shed Statistics - Third Pit Removal)

\n

slope (Slope Gradient and Curvature values)

\n

inverted (Calculating Directions on Inverted DEM)

\n

iwatersheds (Calculating Inverted Watersheds)

\n

ilocal (Initial Inverted Pit Removal)

\n

report (Create the final report)

\n
\n", 
    slug = "details")),2021,list(name = "LITAP", version = "0.4.1.9000", authors = "Steffi LaZerte, Sheng Li, Agriculture and Agri-Food Canada"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "LITAP"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Articles\n
\n
\n  News\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE