Calculate spatial entities required for WEPP

This function takes backup data frame output from flow_mapper() and form_mapper() calculates hillslope, channel segments and impoundment spatial entities required for input in to WEPP (Water Erosion and Prediction Project) (among other metrics). Based on FacetMapR by R. A. (Bob) MacMillan, LandMapper Environmental Solutions.

wepp_mapper(
  folder,
  grid,
  chan_length = 200,
  upslope_threshold = 300,
  out_format = "rds",
  clean = FALSE,
  resume = NULL,
  end = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE
)

Arguments

folder
Character. Location of flow_mapper() output

grid
Numeric. Grid size for the original dem

chan_length
Numeric. Channel length maximum length. Used to split channels into segments

upslope_threshold
Numeric. Threshold of upslope cells to define channel cells. #' Cells with an upslope value larger than this are considered channel cells

out_format
Character. What format should the data be output as? "csv" for Comma-separated values, "dbf" for dbf database files, or "rds" for R data format.

clean
Logical. Remove all output files from previous runs in this folder?

resume
Character. From which stage should the run be resumed? (see Details below)

end
Character. If ending a run after a particular step, which step (see Details below)

log
Logical. Create log file recording progress?

verbose
Logical. Output extra progress messages.

quiet
Logical. Suppress all messages.

Examples

wepp_mapper,Calculate spatial entities required for WEPP,Calculate spatial entities required for WEPP — wepp_mapper,Source: R/wepp_mapper.R,wepp_mapper.Rd,character(0),wepp_mapper,character(0),list(title = "Description", contents = "
This function takes backup data frame output from flow_mapper() and\nform_mapper() calculates hillslope, channel segments and impoundment\nspatial entities required for input in to WEPP (Water Erosion and Prediction\nProject) (among other metrics).  Based on FacetMapR by R. A. (Bob) MacMillan,\nLandMapper Environmental Solutions.
"),list(description = "This function takes backup data frame output from flow_mapper() and\nform_mapper() calculates hillslope, channel segments and impoundment\nspatial entities required for input in to WEPP (Water Erosion and Prediction\nProject) (among other metrics).  Based on FacetMapR by R. A. (Bob) MacMillan,\nLandMapper Environmental Solutions."),wepp_mapper(
  folder,
  grid,
  chan_length = 200,
  upslope_threshold = 300,
  out_format = "rds",
  clean = FALSE,
  resume = NULL,
  end = NULL,
  log = TRUE,
  verbose = FALSE,
  quiet = FALSE
),list(list(name = "folder", description = "
Character. Location of flow_mapper() output
"), list(name = "grid", description = "
Numeric. Grid size for the original dem
"), list(name = "chan_length", description = "
Numeric. Channel length maximum length. Used to split\nchannels into segments
"), list(name = "upslope_threshold", description = "
Numeric. Threshold of upslope cells to define\nchannel cells. #' Cells with an upslope value larger than this are\nconsidered channel cells
"), 
    list(name = "out_format", description = "
Character. What format should the data be output as? \"csv\"\nfor Comma-separated values, \"dbf\" for dbf database files, or \"rds\" for R\ndata format.
"), list(name = "clean", description = "
Logical. Remove all output files from previous runs in this\nfolder?
"), list(name = "resume", description = "
Character. From which stage should the run be resumed? (see\nDetails below)
"), list(name = "end", description = "
Character. If ending a run after a particular step, which step\n(see Details below)
"), 
    list(name = "log", description = "
Logical. Create log file recording progress?
"), list(name = "verbose", description = "
Logical. Output extra progress messages.
"), list(name = "quiet", description = "
Logical. Suppress all messages.
")),TRUE,

if (FALSE) {
# First need to run flow_mapper()
flow_mapper(file = system.file("extdata", "testELEV.dbf", package = "LITAP"),
           out_folder = "./testELEV/", nrow = 90, ncol = 90)

# Now can run wepp_mapper()
wepp_mapper(folder = "./testELEV/", grid = 5)
}

,list(),2021,list(name = "LITAP", version = "0.4.1.9000", authors = "Steffi LaZerte, Sheng Li, Agriculture and Agri-Food Canada"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "LITAP"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Articles\n
\n
\n  News\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE